#1 moved results controllers to separate module
